requirements.txt
asgiref==3.5.2
asttokens==2.2.1
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.12.0
certifi==2022.12.7
click==8.1.3
decorator==5.1.1
Django==4.1.4
djangorestframework==3.14.0
exceptiongroup==1.0.4
executing==1.2.0
h11==0.14.0
idna==3.4
ipdb==0.13.11
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
outcome==1.2.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.6
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.7.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wcwidth==0.2.5
wsproto==1.2.0
